refactor: extract chat tui run state
